OneFormer Type Mismatch Fix
Added two statements in `_pad_image` and `pad` of the OneFormerImageProcessor class.
These guards will take a NoneType input and map it to a float, 0.0.

This has been added since the processor can pass down a NoneType down
into the padding which will result in NoneType errors. Ideally, the
passing in of the NoneType should be remedied instead.